Move GroupByHashMode to top level
