+++ core/trino-main/src/main/java/io/trino/operator/GroupByHashMode.java
@@ -1,0 +1,36 @@
+/*
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package io.trino.operator;
+
+public enum GroupByHashMode {
+    // Hash values are pre-computed as input, and emitted as output
+    PRECOMPUTED,
+    // Hash values are computed by the FlatGroupByHash instance and stored along with the entry
+    CACHED,
+    // Hash values are re-computed for each access on demand
+    ON_DEMAND;
+
+    public boolean isHashPrecomputed()
+    {
+        return this == PRECOMPUTED;
+    }
+
+    public boolean isHashCached()
+    {
+        return switch (this) {
+            case PRECOMPUTED, CACHED -> true;
+            case ON_DEMAND -> false;
+        };
+    }
+}
